Clean more code
Move methods to a separate file.
Extract constants for hardcoded values.
