+++ src/constants.py
@@ -1,0 +1,1 @@
+UNICODE_FOR_A_CHAR = 65
